app/requirements.txt
fastapi
uvicorn
gunicorn
pydantic
torch
torchvision
torchaudio
python-multipart
pillow
requests
aiohttp